fix py tests
