+++ server/tests/static_test.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 import unittest
-from unittest.mock import patch
 
 from main import app
 
@@ -50,15 +49,3 @@ class TestStaticPages(unittest.TestCase):
         response = app.test_client().get('/feedback')
         assert response.status_code == 200
         assert b"We would love to get your feedback!" in response.data
-
-    @patch('routes.static.list_blobs')
-    def test_special_announcement(self, mock_list_blobs):
-        mock_list_blobs.side_effect = (lambda bucket, max_blobs: [])
-        response = app.test_client().get('/special_announcement')
-        assert response.status_code == 200
-        assert b"COVID-19 Special Announcements" in response.data
-
-    def test_special_announcement_faq(self):
-        response = app.test_client().get('/special_announcement/faq')
-        assert response.status_code == 200
-        assert b"COVID-19 Data Feed FAQ" in response.data
